Fixed 'undefined' values for fp, tn, tp, fn:
Evaluation yielded always undefined values for fp, tn, tp and fn.
Simulation name was being changed to nominal during the run of the evaluation
outside of the config file, which resulted in loading the wrong CSV file for
anomalous conditions and thereby not recognizing any collision instances.
Solved the problem with distinct variables for nominal and anomalous simulation names.